樣式調整，剩主題內容頁跟預約資料

+++ themes/theme_list.php
@@ -41,23 +41,19 @@ if ($totalRows) {
 <?php include __DIR__ . '/../parts/html-head.php' ?>
 <?php include __DIR__ . '/../parts/bt-navbar.php' ?>
 
-<div class="container-fluid pt-5">
+<div class="container-fluid p-5">
   <!-- 分頁膠囊   -->
-  <div class="container">
+  <div class="container-fluid">
     <div class="row">
       <!-- 右邊表格 -->
       <div class="col-12">
         <!-- 分頁膠囊 -->
         <ul class="nav nav-pills mb-4" id="pills-tab" role="tablist">
           <li class="nav-item me-3" role="presentation">
-            <button class="nav-link active rounded-pill fw-bold" id="pills-home-tab" data-bs-toggle="pill"
-              data-bs-target="#pills-home" type="button" role="tab" aria-controls="pills-home"
-              aria-selected="true">主題列表</button>
+            <button class="nav-link active rounded-pill fw-bold" id="pills-home-tab" data-bs-toggle="pill" data-bs-target="#pills-home" type="button" role="tab" aria-controls="pills-home" aria-selected="true">主題列表</button>
           </li>
           <li class="nav-item" role="presentation">
-            <button class="nav-link rounded-pill fw-bold" id="pills-profile-tab" data-bs-toggle="pill"
-              data-bs-target="#pills-profile" type="button" role="tab" aria-controls="pills-profile"
-              aria-selected="false">新增主題</button>
+            <button class="nav-link rounded-pill fw-bold" id="pills-profile-tab" data-bs-toggle="pill" data-bs-target="#pills-profile" type="button" role="tab" aria-controls="pills-profile" aria-selected="false">新增主題</button>
           </li>
           <li class="ms-auto">
 
@@ -92,13 +88,13 @@ if ($totalRows) {
                   <th scope="col">時長</th>
                   <th scope="col">開始日期</th>
                   <th scope="col">結束日期</th>
-                  <th scope="col"><i class="fa-solid fa-file-lines"></i></th>
-                  <th scope="col"><i class="fa-solid fa-pen-to-square"></i></th>
-                  <th scope="col"><i class="fa-solid fa-trash"></i></th>
+                  <th scope="col"><i class="fa-solid fa-file-lines"> 檢視</i></th>
+                  <th scope="col"><i class="fa-solid fa-pen-to-square"> 編輯</i></th>
+                  <th scope="col"><i class="fa-solid fa-trash"></i> 刪除</th>
                 </tr>
               </thead>
               <tbody>
-                <?php foreach ($rows as $r): ?>
+                <?php foreach ($rows as $r) : ?>
                   <tr>
                     <td><?= $r['theme_id'] ?></td>
                     <td><?= $r['theme_name'] ?></td>
@@ -108,18 +104,17 @@ if ($totalRows) {
                     <td><?= $r['start_date'] ?></td>
                     <td><?= $r['end_date'] ?></td>
                     <td>
-                      <a href="theme_content.php?theme_id=<?= $r['theme_id'] ?>">
-                        <i class="fa-solid fa-file-lines text-secondary"></i>
+                      <a class="btn btn-dark" href="theme_content.php?theme_id=<?= $r['theme_id'] ?>">
+                        <i class="fa-solid fa-file-lines"> 檢視</i>
                       </a>
                     </td>
                     <td>
-                      <a href="theme_edit.php?theme_id=<?= $r['theme_id'] ?>">
-                        <i class="fa-solid fa-pen-to-square"></i>
+                      <a class="btn btn-primary" href="theme_edit.php?theme_id=<?= $r['theme_id'] ?>">
+                        <i class="fa-solid fa-pen-to-square"> 編輯</i>
                       </a>
                     </td>
-                    <td><a href="theme_delete.php?theme_id=<?= $r['theme_id'] ?>"
-                        onclick="return confirm('是否要刪除編號為<?= $r['theme_id'] ?>的資料')">
-                        <i class="fa-solid fa-trash text-danger"></i>
+                    <td><a class="btn btn-danger" href="theme_delete.php?theme_id=<?= $r['theme_id'] ?>" onclick="return confirm('是否要刪除編號為<?= $r['theme_id'] ?>的資料')">
+                        <i class="fa-solid fa-trash text-white"> 刪除</i>
                       </a></td>
                   </tr>
                 <?php endforeach; ?>
@@ -139,12 +134,12 @@ if ($totalRows) {
                       <i class="fa-solid fa-angle-left"></i>
                     </a>
                   </li>
-                  <?php for ($i = $page - 5; $i <= $page + 5; $i++):
-                    if ($i >= 1 and $i <= $totalPages): ?>
+                  <?php for ($i = $page - 5; $i <= $page + 5; $i++) :
+                    if ($i >= 1 and $i <= $totalPages) : ?>
                       <li class="page-item <?= $page == $i ? 'active' : '' ?> ">
                         <a class="page-link" href="?page=<?= $i ?>"><?= $i ?></a>
                       </li>
-                    <?php endif;
+                  <?php endif;
                   endfor ?>
                   <li class="page-item">
                     <a class="page-link" href="#">
@@ -173,7 +168,7 @@ if ($totalRows) {
 <?php include __DIR__ . '/../parts/scripts.php' ?>
 
 <script>
-  document.getElementById('searchForm').addEventListener('submit', function (event) {
+  document.getElementById('searchForm').addEventListener('submit', function(event) {
     event.preventDefault(); // 阻止表單提交
 
     var formData = new FormData(this);
